refactor(logger-middleware): refactor into proper middleware module
